--- engine/app/components/citizens_advice_components/checkable/checkbox.rb
@@ -1,11 +1,0 @@
-# frozen_string_literal: true
-
-module CitizensAdviceComponents
-  module Checkable
-    class Checkbox < Base
-      def input_type
-        "checkbox"
-      end
-    end
-  end
-end
