--- engine/app/components/citizens_advice_components/checkable/radio.rb
@@ -1,11 +1,0 @@
-# frozen_string_literal: true
-
-module CitizensAdviceComponents
-  module Checkable
-    class Radio < Base
-      def input_type
-        "radio"
-      end
-    end
-  end
-end
