+++ engine/app/components/citizens_advice_components/legacy/checkable/checkbox.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+module CitizensAdviceComponents
+  module Legacy
+    module Checkable
+      class Checkbox < Base
+        def input_type
+          "checkbox"
+        end
+      end
+    end
+  end
+end
