+++ engine/app/components/citizens_advice_components/legacy/checkable/radio.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+module CitizensAdviceComponents
+  module Legacy
+    module Checkable
+      class Radio < Base
+        def input_type
+          "radio"
+        end
+      end
+    end
+  end
+end
