--- engine/app/components/citizens_advice_components/text_input.rb
@@ -1,54 +1,0 @@
-# frozen_string_literal: true
-
-module CitizensAdviceComponents
-  class TextInput < Input
-    attr_reader :base_input_args
-
-    def initialize(width: nil, **args)
-      @width = fetch_or_fallback(
-        allowed_values: allowed_width_values,
-        given_value: width,
-        fallback: nil
-      )
-      args[:options]&.merge(width: @width)
-      @base_input_args = args
-      super(**@base_input_args)
-    end
-
-    def call
-      render CitizensAdviceComponents::Input.new(**base_input_args) do
-        tag.input(**input_attributes)
-      end
-    end
-
-    def allowed_width_values
-      [nil, :two_chars, :four_chars, :eight_chars, :sixteen_chars]
-    end
-
-    def allowed_type_values
-      %i[
-        email
-        number
-        password
-        search
-        tel
-        text
-        url
-      ]
-    end
-
-    def width?
-      @width.present?
-    end
-
-    def class_names
-      classes = ["cads-input"]
-      classes << "cads-input--#{@width.to_s.dasherize}" if width?
-      classes.join(" ")
-    end
-
-    def base_input_attributes
-      super.merge(class: class_names)
-    end
-  end
-end
